+++ test/index.js
@@ -1,0 +1,7 @@
+'use strict'
+
+const Path = require('path')
+const Dotenv = require('dotenv')
+
+// import environment variables from local secrets.env file
+Dotenv.config({ path: Path.resolve(__dirname, 'secrets.env') })